requirements.txt
certifi~=2019.9.11
cffi~=1.13.2
chardet~=3.0.4
cryptography~=2.8
defusedxml~=0.6.0
idna~=2.7
jira~=2.0.0
notify-run~=0.0.12
oauthlib~=3.1.0
pbr~=5.4.4
pycparser~=2.19
PyJWT~=1.7.1
PyQRCode~=1.2.1
python-dateutil~=2.8.1
python-gitlab~=1.13.0
requests~=2.20.0
requests-oauthlib~=1.3.0
requests-toolbelt~=0.9.1
six~=1.13.0
urllib3~=1.24.3
